+++ src/frontend/apps/desk/src/features/teams/api/index.ts
@@ -1,2 +1,3 @@
 export * from './useCreateTeam';
 export * from './useTeam';
+export * from './useTeams';
